Expose additional session variables, add more tests, ensure Cbor is turned off.
Added new session variable iter_start_timestamp to start iterator from a specific place, and made
the message timestamp a Presto timestamp field.  Exposed a few other session variables.
Added new unit test to ensure we are giving
the session variables to the connector manager and that we can get/set the values.